refactor: remove unnessasary code, replace some code
